# contracts/openapi.yaml
openapi: "3.0.0"
info:
  version: "0.1.0"
  title: ContextPilot API
  description: |
    Meeting intelligence REST API.
    Phase 0 placeholder: metadata and common schemas only.
    Endpoints added in BRD-02 and later domain specs.
servers:
  - url: /api/v1
    description: Base API path
...
security:
  - BearerAuth: []
  - CookieAuth: []
paths:
...
  /upcoming/{meetingId}/briefing:
    get:
      operationId: getActiveBriefing
      summary: Active/latest briefing version for an upcoming meeting
      description: Returns the active (latest completed) briefing version for an upcoming meeting. Gated by FF_ENABLE_PRE_CALL_BRIEFING.
      tags:
        - briefing
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active briefing version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingVersion'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled or no briefing available
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Upcoming meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /upcoming/{meetingId}/briefing/versions:
    get:
      operationId: listBriefingVersions
      summary: Version history list
      description: Returns all briefing versions for an upcoming meeting, ordered by version number descending.
      tags:
        - briefing
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Briefing version history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingVersionList'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /upcoming/{meetingId}/briefing/versions/{versionNumber}:
    get:
      operationId: getBriefingVersion
      summary: Specific briefing version
      description: Returns a specific briefing version by version number.
      tags:
        - briefing
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: versionNumber
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Specific briefing version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingVersion'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Upcoming meeting or version not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /upcoming/{meetingId}/briefing/regenerate:
    post:
      operationId: regenerateBriefing
      summary: Manual regeneration request
      description: Queues an async briefing regeneration. Latest completed briefing remains available during regeneration. Gated by FF_ENABLE_PRE_CALL_BRIEFING.
      tags:
        - briefing
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '202':
          description: Regeneration job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingRegenerateAccepted'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '409':
          description: Regeneration already in progress
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /upcoming/{meetingId}/briefing/sources/{sourceId}/exclude:
    post:
      operationId: excludeBriefingSource
      summary: Exclude a source meeting
      description: Marks a source meeting as excluded for the given upcoming meeting. Exclusions persist for future regenerations. Gated by FF_ENABLE_PRE_CALL_BRIEFING.
      tags:
        - briefing
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Source excluded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingSourceExclusion'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled or source not accessible
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Upcoming meeting or source meeting not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
  /upcoming/{meetingId}/briefing/sources/{sourceId}/restore:
    post:
      operationId: restoreBriefingSource
      summary: Restore an excluded source meeting
      description: Restores (undoes) a source exclusion for the upcoming meeting. The exclusion record is soft-deleted via restored_at. Gated by FF_ENABLE_PRE_CALL_BRIEFING.
      tags:
        - briefing
      security:
        - BearerAuth: []
        - CookieAuth: []
      parameters:
        - name: meetingId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: sourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Source restored
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BriefingSourceExclusion'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '403':
          description: Feature flag disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: Exclusion record not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
...
  schemas:
    Problem:
      description: RFC 7807 error response
      type: object
      properties:
        type:
          type: string
          format: uri
          example: https://contextpilot.app/errors/problem
        title:
          type: string
          example: Resource Not Found
        status:
          type: integer
          example: 404
        detail:
          type: string
          example: Meeting with ID 12345 does not exist or you do not have access.
        instance:
          type: string
          format: uri
          example: /api/v1/meetings/12345
        request_id:
          type: string
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    CookieAuth:
      type: apiKey
      in: cookie
      name: session

  # ─── Memory Processing Schemas ───────────────────────────────────────────

    MemoryProcessingState:
      type: object
      description: Current processing lifecycle state for a meeting's memory.
      properties:
        meeting_id:
          type: string
          format: uuid
        processing_state:
          type: string
          enum: [not_processed, queued, processing, completed, completed_with_insufficient_evidence, failed, stale, reprocessing, retrying, retry_exhausted]
        active_version_number:
          type: integer
          nullable: true
          description: Present when a memory version exists.
        is_stale:
          type: boolean
          description: True when source meeting has been edited after the active version was created.
        latest_job:
          type: object
          nullable: true
          properties:
            job_id:
              type: string
              format: uuid
            status:
              type: string
              enum: [queued, processing, completed, completed_with_insufficient_evidence, failed, retrying, retry_exhausted]
            retry_count:
              type: integer
...
  securitySchemes: